% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{creer_boucle_rectangle}
\alias{creer_boucle_rectangle}
\title{Fonction de secours pour créer un rectangle si on arrive pas a créer une boucle plus compliqué
Générer une boucle rectangulaire simple}
\usage{
creer_boucle_rectangle(n)
}
\arguments{
\item{n}{Entier. Taille de la grille.}
}
\value{
Une liste contenant les segments, les points visités et la taille \code{n}.
}
\description{
Fonction de secours utilisée lorsque la génération aléatoire échoue.
Elle construit une boucle rectangulaire qui fait le tour complet de la grille.
}
\examples{
boucle <- creer_boucle_rectangle(4)
}
